Make fetch_groups use a separate admin service account
The former way how fetch_groups worked wasn't much usable. It used the
user token to look up the groups in the admin directory API. This means
all users need to have access to the admin API which is usually a bad
idea.

This new approach uses a separate service account for the group lookups
so the authenticating users don't need any permission on that API.